Tap Master
Developed a game for iOS and Android platforms. This game tests the reaction time of the user. Random objects of different shapes and different colors are generated and moved in random direction during random time intervals similar to a popular game Fruit Ninja. The user only has to tap some specific shapes of a specific color. The score is generated based on the taps of the user. Used Lua as the programming language and Corona SDK for development of the project.